updated markdown to better format
the ")" after the 1. and 2. was unnecessary and prevented the program from making it a list. You can also make a list by just putting a dash: 

- one
- two

One thought about turning the prompt into a more "DH-y" project. Have the students collect their questions, the transcriptions of the answers, and all the data about the interviewees in a structured format -- perhaps just in a spreadsheet, but we could also throw up a basic site to pull them from some blank data. 

One other thought -- would it be interesting to collaborate with your class to come up with the same set of seven questions? That way, the answers could be compared later in order to elicit commonalities and differences ...